third_party/WebKit/LayoutTests/external/wpt/cookies/samesite/form-post-blank-expected.txt

This is a testharness.js-based test.
PASS Same-host top-level form POSTs are strictly same-site
PASS Subdomain top-level form POSTs are strictly same-site
FAIL Cross-site top-level form POSTs are cross-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.9190326941237728" but got (undefined) undefined
PASS Same-host redirecting to same-host top-level form POSTs are strictly same-site
PASS Subdomain redirecting to same-host top-level form POSTs are strictly same-site
PASS Cross-site redirecting to same-host top-level form POSTs are strictly same-site
PASS Same-host redirecting to subdomain top-level form POSTs are strictly same-site
PASS Subdomain redirecting to subdomain top-level form POSTs are strictly same-site
PASS Cross-site redirecting to subdomain top-level form POSTs are strictly same-site
FAIL Same-host redirecting to cross-site top-level form POSTs are cross-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.6502580520675447" but got (undefined) undefined
FAIL Subdomain redirecting to cross-site top-level form POSTs are cross-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.5102504378195483" but got (undefined) undefined
FAIL Cross-site redirecting to cross-site top-level form POSTs are cross-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.737020387720801" but got (undefined) undefined
Harness: the test ran to completion.
